+++ oocana/oocana/context.py
@@ -2,7 +2,7 @@ import asyncio
 from dataclasses import asdict
 from .data import BlockInfo, StoreKey, JobDict, BlockDict, BinValueDict, VarValueDict
 from .mainframe import Mainframe
-from .handle_data import HandleDef, OutputHandleDef
+from .handle import HandleDef, OutputHandleDef
 from typing import Dict, Any, TypedDict, Optional, Callable, Mapping
 from types import MappingProxyType
 from base64 import b64encode
