+++ oocana/tests/test_handle.py
@@ -1,5 +1,5 @@
 import unittest
-from oocana import handle_data
+from oocana.handle import HandleDef, InputHandleDef, FieldSchema
 from typing import cast
 
 
@@ -15,23 +15,23 @@ class TestHandleData(unittest.TestCase):
     def test_handle_def_field(self):
 
         d = fixture.copy()
-        handle_data.HandleDef(**d)
+        HandleDef(**d)
 
         d = fixture.copy()
         del d["name"]
 
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertEqual(handle_def.handle, "test")
         self.assertIsNotNone(handle_def.json_schema)
         
-        json_schema = cast(handle_data.FieldSchema, handle_def.json_schema)
+        json_schema = cast(FieldSchema, handle_def.json_schema)
         self.assertEqual(json_schema.contentMediaType, "oomol/bin")
 
         d = fixture.copy()
         del d["name"]
         del d["json_schema"]
 
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertEqual(handle_def.handle, "test")
         self.assertIsNone(handle_def.json_schema)
 
@@ -39,7 +39,7 @@ class TestHandleData(unittest.TestCase):
         d = fixture.copy()
         d["a"] = "a"
 
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertEqual(handle_def.handle, "test")
 
     def test_handle_def_missing_field(self):
@@ -48,17 +48,17 @@ class TestHandleData(unittest.TestCase):
         }
 
         with self.assertRaises(ValueError, msg="missing attr key: 'handle'"):
-            handle_data.InputHandleDef(**d) # type: ignore
+            InputHandleDef(**d) # type: ignore
 
     def test_handle_type(self):
         d = fixture.copy()
         d["json_schema"]["contentMediaType"] = "oomol/secret"
 
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertTrue(handle_def.is_secret_handle())
 
         d["json_schema"]["contentMediaType"] = "oomol/var"
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertTrue(handle_def.is_var_handle())
 
         d = {
@@ -75,7 +75,7 @@ class TestHandleData(unittest.TestCase):
                 "type": "array"
             }
         }
-        handle_def = handle_data.HandleDef(**d)
+        handle_def = HandleDef(**d)
         self.assertFalse(handle_def.is_var_handle())
         self.assertFalse(handle_def.is_secret_handle())
 
@@ -83,11 +83,11 @@ class TestHandleData(unittest.TestCase):
         d = fixture.copy()
         d["json_schema"]["contentMediaType"] = "oomol/secret"
 
-        handle_def = handle_data.InputHandleDef(**d)
+        handle_def = InputHandleDef(**d)
         self.assertTrue(handle_def.is_secret_handle())
 
         d["json_schema"]["contentMediaType"] = "oomol/var"
-        handle_def = handle_data.InputHandleDef(**d)
+        handle_def = InputHandleDef(**d)
         self.assertTrue(handle_def.is_var_handle())
 
         d = {
@@ -104,6 +104,6 @@ class TestHandleData(unittest.TestCase):
                 "type": "array"
             }
         }
-        handle_def = handle_data.InputHandleDef(**d)
+        handle_def = InputHandleDef(**d)
         self.assertFalse(handle_def.is_var_handle())
         self.assertFalse(handle_def.is_secret_handle())
